refactor: student model as a distinct app
