+++ backend/backend/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'base',
     'internship',
     'application',
+    'student',
 ]
 
 MIDDLEWARE = [
