+++ backend/backend/urls.py
@@ -20,6 +20,7 @@ from django.contrib.auth import urls
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('v1/api/', include('base.urls')),
+    path('v1/api/student/', include('student.urls')),
     path('v1/api/internship/', include('internship.urls')),
     path('v1/api/application/', include('application.urls')),
 ]
