+++ backend/base/admin.py
@@ -3,5 +3,5 @@ from .models import *
 # Register your models here.
 admin.site.register(User)
 admin.site.register(Company)
-admin.site.register(Student)
+
 # Register your models here.
