+++ backend/base/models.py
@@ -1,5 +1,3 @@
-from email.errors import NonPrintableDefect
-from tkinter import CASCADE
 from django.db import models
 from django.contrib.auth.models import AbstractUser
 from django.db.models.signals import post_save
@@ -39,18 +37,7 @@ class Company(models.Model):
         return self.user.email
 
 
-class Student(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    sid = models.AutoField(unique=True, primary_key=True)
-    firstName = models.CharField(max_length=200, null=True, blank=True)
-    lastName = models.CharField(max_length=200, null=True, blank=True)
-    universityName = models.CharField(default="",max_length=100, null=True, blank=True)
-    major = models.CharField(default="",max_length=100, null=True, blank=True)
-    gender = models.CharField(max_length=200, null=True, blank=True)
-    cGpa = models.DecimalField(max_digits=3, decimal_places=2, null=True, blank=True)
-    img = models.CharField(max_length=1000, null=True, blank=True)
-    def __str__(self):
-        return self.user.email
+
 
 
 # @receiver(post_save, sender=User)
