+++ backend/base/permissions.py
@@ -13,13 +13,6 @@ class CanManageInternships(BasePermission):
             return True
         return False
 
-
-class IsStudentUser(BasePermission):
-    def has_permission(self, request, view):
-
-        return bool(request.user and request.user.is_student)
-
-
 class IsCompanyUser(BasePermission):
     def has_permission(self, request, view):
