+++ backend/base/urls.py
@@ -4,20 +4,13 @@ urlpatterns = [
     path("users/", views.UserListCreateView.as_view(), name="all-users"),
     path('users/<int:pk>/', views.UserDetailView.as_view, name='user-update-delete-get'),
 
-    path('signup/company/', views.CompanySignupView.as_view()),
-    path('signup/student/', views.StudentSignupView.as_view()),
-
     path('login/', views.CustomAuthToken.as_view(), name='auth-token'),
     path('logout/', views.LogoutView.as_view(), name='logout-view'),
 
+    path('signup/company/', views.CompanySignupView.as_view()),
     path('company/dashboard/', views.CompanyOnlyView.as_view(), name='company-dashboard'),
-    path('student/dashboard/', views.StudentOnlyView.as_view(), name='student-dashboard'),
-
-    path('student/<int:pk>/', views.StudentDetailView.as_view(), name='student-verify'),
-    path('student/all/', views.StudentListView.as_view(), name='all-students'),
-
-    path('student/edit/<int:pk>/', views.StudentProfileAPIView.as_view(), name='student-get-or-update'),
-    path('company/', views.CompanyProfileAPIView.as_view(), name='company-get-or-update'),
+    path('company/<int:pk>/', views.CompanyDetailView.as_view(), name='get-company-data-from-userid'),
+    path('company/<int:pk>/', views.CompanyProfileAPIView.as_view(), name='company-get-or-update'),
 
     
 ]
