+++ backend/student/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import *
+# Register your models here.
+admin.site.register(Student)
